requirements.txt.txt
customtkinter
pycryptodome